Create Travelling.c
The Traveling Salesman Problem (TSP) is a well-known combinatorial optimization problem. Solving TSP for large sets of cities is computationally intensive, and an optimal solution may not be feasible in a reasonable amount of time. However, you can use various algorithms to find approximate solutions. One such algorithm is the Nearest Neighbor Algorithm.